refactor imports to use barrel files for ErrorPageComponent and authentication pages
