+++ src/app/app-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { ErrorPageComponent } from './presentation/pages/error-page/error-page.component';
+import { ErrorPageComponent } from './presentation/pages';
+
 
 //*---------------------------------------------------------
 //* Routes
