+++ src/app/modules/auth/auth.module.ts
@@ -1,19 +1,24 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AuthLayoutComponent } from '../../presentation/layouts/auth-layout/auth-layout.component';
+import { LoginPageComponent, SignupPageComponent } from '../../presentation/pages';
 import { AuthRoutingModule } from '../routes/auth-routing.module';
-import { LoginPageComponent } from '../../presentation/pages/login-page/login-page.component';
-import { SignupPageComponent } from '../../presentation/pages/signup-page/signup-page.component';
+
+
 
 
 
 
 @NgModule({
   declarations: [
+    //? Layouts
     AuthLayoutComponent,
 
+    //? Pages
     LoginPageComponent,
-    SignupPageComponent
+    SignupPageComponent,
+
+    //? Components
   ],
   imports: [
     CommonModule,
